+++ examples/distro/buildSrc/src/main/kotlin/Versions.kt
@@ -1,8 +1,8 @@
 // this line is managed by .github/scripts/update-sdk-version.sh
-const val opentelemetrySdkVersion = "1.64.0"
+const val opentelemetrySdkVersion = "1.65.0"
 
 // these lines are managed by .github/scripts/update-version.sh
-const val opentelemetryJavaagentVersion = "2.30.2"
-const val opentelemetryJavaagentAlphaVersion = "2.30.2-alpha"
+const val opentelemetryJavaagentVersion = "2.31.1"
+const val opentelemetryJavaagentAlphaVersion = "2.31.1-alpha"
 
 const val autoserviceVersion = "1.1.1"