+++ instrumentation/spring/spring-boot-autoconfigure/testing/src/main/java/io/opentelemetry/instrumentation/spring/autoconfigure/internal/AbstractJdbcInstrumentationAutoConfigurationTest.java
@@ -82,8 +82,7 @@ public abstract class AbstractJdbcInstrumentationAutoConfigurationTest {
                           trace.hasSpansSatisfyingExactly(
                               span ->
                                   span.hasAttribute(maybeStable(DB_STATEMENT), "SELECT ?")
-                                      .hasAttribute(
-                                          DB_QUERY_PARAMETER.getAttributeKey("0"), "1")));
+                                      .hasAttribute(DB_QUERY_PARAMETER.getAttributeKey("0"), "1")));
             });
   }
 }